refactor: Moved pages to pages dir
