+++ src/App.tsx
@@ -2,8 +2,8 @@ import './styles/app.scss';
 import './styles/globals.scss';
 import { NavBar } from './components/Organisms/NavBar';
 import { Routes, Route } from 'react-router-dom';
-import { Home } from './components/Templates/Home';
-import Categories from './components/Templates/Categories/Categories.component';
+import { Home } from './components/Templates/Pages/Home';
+import { Categories } from './components/Templates/Pages/Categories';
 
 function App() {
   return (
