+++ src/components/Templates/Pages/Categories/index.tsx
@@ -1,0 +1,1 @@
+export { default as Categories } from './Categories.component'
